Pass necessary objects into create MCP server code
