+++ src/core/prompts/instructions/create-mode.ts
@@ -1,3 +1,3 @@
-export function createModeInstructions(): string {
+export async function createModeInstructions(): Promise<string> {
 	return "This is how to create a new mode"
 }
